fix: consolidate legacy and default plugin and fix resizing (#2254)
* fix: consolidate legacy and default plugin and fix ER/EV resizing

* fix: show offline message on EE maps

---------
