--- src/components/Item/VisualizationItem/Visualization/MapPlugin.js
@@ -1,30 +1,0 @@
-import { useDhis2ConnectionStatus } from '@dhis2/app-runtime'
-import i18n from '@dhis2/d2-i18n'
-import PropTypes from 'prop-types'
-import React from 'react'
-import IframePlugin from './IframePlugin.js'
-import NoVisualizationMessage from './NoVisualizationMessage.js'
-
-const mapViewIsEELayer = (mapView) => mapView.layer.includes('earthEngine')
-
-const MapPlugin = ({ visualization, ...pluginProps }) => {
-    const { isDisconnected: offline } = useDhis2ConnectionStatus()
-
-    if (offline && visualization.mapViews?.find(mapViewIsEELayer)) {
-        return (
-            <NoVisualizationMessage
-                message={i18n.t(
-                    'Maps with Earth Engine layers cannot be displayed when offline'
-                )}
-            />
-        )
-    }
-
-    return <IframePlugin visualization={visualization} {...pluginProps} />
-}
-
-MapPlugin.propTypes = {
-    visualization: PropTypes.object,
-}
-
-export default MapPlugin
